requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
torch==2.6.0
torchvision==0.21.0
Pillow==11.2.1
numpy==2.0.2
python-multipart==0.0.20
requests==2.32.3
pydantic==2.11.3
torchserve==0.12.0
torch-model-archiver==0.12.0